Added 2 new internal features. Changed 2 lines in channelDelete file
Voice channel creation now check a user is present before creating a channel, this is a 500 millisecond delay.

Users are given an internal value which results in a cooldown for 15 seconds, which prevents a user from creating a new voice channel in that time frame.

Changed bracket spacing for channelDelete file.