Added new Feature, Error management and Fixed Kick command
Added a channel bypass for Managers and Staff members added to the bot (This also applies to roles), they can now join locked channels. Hooray! \o/

Fixed Kick command, when a channel owner kicked a member from the channel, it would result in a loss of ownership of the voice channel.

Added Error Management for every Message and React occurrence.