New Feature! + Complete rework of the voiceStateUpdate
Rework of the voiceStateUpdate file, this now contains a new feature. Discord only allows 50 channels per category, this includes text channels, the bot now features a new overFlow system, Sounds techy :P, posh word for new category when the main one is nearing full, the new category is only a temporary one, it gets removed on the last channel to be removed. The new category is created once every 30 channels, for a totally of 2 extra categories, making for a total of 88 Voice channels.

Edited channelDelete file for new overFlow deletion on removal.

Edited Clear command to accommodate the new overFlow feature and now clears naughty categories >:F

Added new compaction event for dbGuilds file, as this is where the OVERFLOW categories are saved.

Added entries in en lang file for new features and edits.